worker/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
transformers
peft
torch
pika
prometheus_client